+++ vault/resource_aws_secret_backend.go
@@ -151,6 +151,12 @@ func awsSecretBackendResource() *schema.Resource {
 				Computed:    true,
 				Description: "The TTL of generated identity tokens in seconds.",
 			},
+			consts.FieldMaxRetries: {
+				Type:        schema.TypeInt,
+				Optional:    true,
+				Computed:    true,
+				Description: "Number of max retries the client should use for recoverable errors.",
+			},
 		},
 	}, false)
 
@@ -264,6 +270,10 @@ func awsSecretBackendCreate(ctx context.Context, d *schema.ResourceData, meta in
 		}
 	}
 
+	if v, ok := d.GetOk(consts.FieldMaxRetries); ok && v != 0 {
+		data[consts.FieldMaxRetries] = v.(int)
+	}
+
 	if region != "" {
 		data[consts.FieldRegion] = region
 	}
@@ -370,6 +380,10 @@ func awsSecretBackendRead(ctx context.Context, d *schema.ResourceData, meta inte
 		}
 	}
 
+	if err := d.Set(consts.FieldMaxRetries, resp.Data[consts.FieldMaxRetries]); err != nil {
+		return diag.Errorf("error reading %s for AWS Secret Backend %q: %q", consts.FieldMaxRetries, path, err)
+	}
+
 	if err := d.Set(consts.FieldPath, path); err != nil {
 		return diag.FromErr(err)
 	}
@@ -400,7 +414,7 @@ func awsSecretBackendUpdate(ctx context.Context, d *schema.ResourceData, meta in
 	if d.HasChanges(consts.FieldAccessKey,
 		consts.FieldSecretKey, consts.FieldRegion, consts.FieldIAMEndpoint,
 		consts.FieldSTSEndpoint, consts.FieldSTSFallbackEndpoints, consts.FieldSTSRegion, consts.FieldSTSFallbackRegions,
-		consts.FieldIdentityTokenTTL, consts.FieldIdentityTokenAudience, consts.FieldRoleArn,
+		consts.FieldIdentityTokenTTL, consts.FieldIdentityTokenAudience, consts.FieldRoleArn, consts.FieldMaxRetries,
 		consts.FieldRotationSchedule,
 		consts.FieldRotationPeriod,
 		consts.FieldRotationWindow,
@@ -450,6 +464,11 @@ func awsSecretBackendUpdate(ctx context.Context, d *schema.ResourceData, meta in
 			if identityTokenTTL != 0 {
 				data[consts.FieldIdentityTokenTTL] = identityTokenTTL
 			}
+		}
+
+		maxRetries := d.Get(consts.FieldMaxRetries).(int)
+		if maxRetries != 0 {
+			data[consts.FieldMaxRetries] = maxRetries
 		}
 
 		region := d.Get(consts.FieldRegion).(string)